Move poll functions to their own dedicated `.py` file, instead of per nodetree definition file
