+++ utils.py
@@ -91,18 +91,6 @@ def update_tally(context, entry):
         json.dump(tally_dict, f, indent=4)
 
 
-def check_mix_color_alias(_context, valid_options):
-    current_pref = fetch_user_prefs("mix_color_alias")
-    return current_pref in valid_options
-
-
-def in_nodegroup(context):
-    current_tree = context.space_data.edit_tree
-    node_groups = context.blend_data.node_groups
-
-    return current_tree in node_groups.values()
-    
-
 def fetch_active_nodetree(context):
     edit_tree = context.space_data.edit_tree
     node_tree = context.space_data.node_tree
